+++ src/libraries/PID/DEventRFBunch_factory.cc
@@ -21,7 +21,7 @@ jerror_t DEventRFBunch_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DEventRFBunch_factory::brun(jana::JEventLoop *locEventLoop, int runnumber)
+jerror_t DEventRFBunch_factory::brun(jana::JEventLoop *locEventLoop, int32_t runnumber)
 {
 	DApplication* locApplication = dynamic_cast<DApplication*>(locEventLoop->GetJApplication());
 	DGeometry* locGeometry = locApplication->GetDGeometry(runnumber);
@@ -42,7 +42,7 @@ jerror_t DEventRFBunch_factory::brun(jana::JEventLoop *locEventLoop, int runnumb
 //------------------
 // evnt
 //------------------
-jerror_t DEventRFBunch_factory::evnt(JEventLoop* locEventLoop, int eventnumber)
+jerror_t DEventRFBunch_factory::evnt(JEventLoop* locEventLoop, uint64_t eventnumber)
 {
 	//There should ALWAYS be one and only one DEventRFBunch created.
 		//If there is not enough information, time is set to NaN
